Project 7 input

+++ xinu-hw6/include/kernel.h
@@ -25,7 +25,7 @@ syscall kgetc(void);
 syscall kcheckc(void);
 syscall kvprintf(const char *fmt, va_list va);
 
-syscall create(void *, ulong, ulong, char *, ulong, ...);
+syscall create(void *, ulong, uint, char *, ulong, ...);
 syscall kill(pid_typ);
 syscall ready(pid_typ, bool);
 syscall resched(void);
